Functions!
Wrapping the functionality in a class seemed forced.
