+++ .travis.yml
@@ -1,10 +1,9 @@
 language: php
  
 php:
-  - 5.3
-  - 5.4
   - 5.5
   - 5.6
+  - 7.0
   - hhvm
  
 before_script:
