+++ composer.json
@@ -11,12 +11,11 @@
     ],
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.6.0",
+        "beberlei/assert": "^2.5"
     },
     "autoload": {
-        "psr-4": {
-            "Viite\\": "src"
-        }
+        "files": ["src/viite.php"]
     },
     "require-dev": {
         "phpunit/phpunit": "4.7.*"
